ARUHA-56: working on oauth test
